Update gradle, kotlin, and android support libs
This produces a warning about kotlin runtime JAR files in the classpath
having different versions, likely because the kpreferences library is
using an older version (thus, I need to update that library).

Gradle was intentionally not updated to the absolute most recent
version, because it is not yet compatible with the latest stable version
of Android Studio (3.3).