Suppress ExpiredTargetSdkVersion lint error (temp)
The Play Store requires app to target sdk 26 at minimum. This is causing
a lint error that makes Travis fail. We're not ready to target sdk 26
yet (it requires some changes with the SYSTEM_OVERLAY window type), and
Red Moon's Play Store listing has some issues anyway, so it's ok to
suppress the error for now; it doesn't affect F-Droid.