Don't indent preferences in new androidx style
1. Although the design of Red Moon uses preferences to build its UI,
   the interaction patterns are more like the main UI than preferences.
2. Although this new indent thing is the new standard for Android
   preferences, it looks like trash.

There is something to be said for not fighting platform defaults. And.
There is something to be said for not degrading the appearance of an app
  in a point release, which I hope the next one turns out to be.

So, no updated style for now. I do have an idea for a slightly updated
interface that will have non-filter preferences in a dedicated screen.
If I get around to implementing that, I'll reconsider the new style.