Remove MIT license for clarity
Red Moon is a derivative of the MIT-licensed Shades, and some bits have
been contributed under the MIT license. However, to the extent of my
knowledge every file derived from MIT-licensed code has a notice to that
effect in a comment at the top. Having multiple license notices makes an
already confusing topic more unclear, so if it's not needed, it's better
to just provide a single license file. Red Moon is GPL (v3.0+).

Further, the MIT license does not forbid redistributing under the GPL as
long as credit is given, which is has been. So there's really no reason
to continue distributing those bits under the original license except to
respect the intent of the original authors. But at this point any MIT
code has been modified enough that nobody should try to selectively use
parts of Red Moon under the MIT license anyway.