Move non-filter settings to dedicated screen
Fixes #264

A secondary goal here is to rewrite the UI in a more.. idiomatic?.. way.
Much of Red Moon is written in ways that work, but aren't necessarily
the "proper" way to do things on Android. For example, 100% of the UI is
using Preferences. I'm sure that made it fast to write, but it has been
difficult to develop on top of. Also, it meant I never really learned
some basics of building UI on Android, and I'd like to do that.

One technical consequence here is that I've left in a bit of incidental
duplication (eg, ThemedAppCompatActivity and SettingsActivity share a
lot of code). That's because I think I will likely rewrite more of the
UI in the future, so I don't want to tie the settings activity to the
main one until it's clear where the duplication will be.

I had a bit of poor git hygiene here. I worked on a lot of pieces in
parallel, so I never came to a good intermediate place to commit (but
could have if I'd paid more attention to it). If this were more of a
team project (ie, if anyone were going to review this) I'd probably have
gone back and broken it up, but as it stands, there's not much of a
point to that. But, because I'm idealistic, I've still tried to stop at
the earlier point I could. That is to say, there's still some cleanup
work to follow ^_^