Disable filter prefs pending overlay permission
Red Moon is kind of useless if you don't grant the overlay permission.
So, we want people to grant this permission as early as possible
(without denying access to the app). It's also kind of confusing if you
move a slider and it causes the permission prompt to pop up (since the
filter tries to turn on, to show the preview).

Now we have the FAB back again, there's a pretty clear first action for
people to take, so we don't need to worry that they might be confused
about what to do if the filter prefs are disabled.