Allow setting sunrise and sunset individually
This is not quite ready to merge as I'd like to update a few strings to
use String.format instead of the ad-hoc substitution (or worse) that
we're using now.

It's also an initial implementation, where quite a bit is done in the
SettingsFragment. As I continue to break the dialog creation component
out into a library, it's possible that some functionality will migrate
with it.