Implement TimePreferenceDialogFragmentCompat
Okay, so it was still a large commit here because I forgot to commit
midway through. Oops.

There is one outstanding bug: The neutral button is currently not very
useful if you don't set a custom listener, because the default listener
just sets the time picker back to what it was when the dialog opened.
Not very useful considering you can just close the dialog. If anyone
else ever uses the library, we can work on making it more useful :P